meeting_assistant/requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
python-dotenv==1.0.0
whisper-openai==0.1.0
pydub==0.25.1

